+++ packages/runtime/package.json
@@ -1,15 +1,19 @@
 {
   "name": "@bricking/runtime",
   "version": "1.0.0",
-  "description": "",
+  "description": "bricking 运行时",
   "main": "index.js",
   "scripts": {
-    "build": "tsc "
+    "build": "run-s prepare && tsc",
+    "prepare": "del dist && cpy 'src/global.d.ts' 'package.json' dist"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@types/systemjs": "^6.1.1",
     "systemjs": "^6.12.1"
+  },
+  "devDependencies": {
+    "typescript": "^4.5.5"
   }
 }
